+++ lib/Interpreter/SQL/MySQLAggregateOptimizer.php
@@ -284,7 +284,7 @@ class MySQLAggregateOptimizer extends MySQLOptimizer {
 					')' .
 				') * 1000';
 
-		$this->aggFrom = (double) $this->conn->fetchColumn($sql, $sqlParameters, 0);
+		$this->aggFrom = (float) $this->conn->fetchColumn($sql, $sqlParameters, 0);
 		$this->aggTo = null;
 
 		// aggregate table contains relevant data?
@@ -301,7 +301,7 @@ class MySQLAggregateOptimizer extends MySQLOptimizer {
 				$sqlParameters[] = $this->to;
 				$sql .= ' AND timestamp<?';
 			}
-			$this->aggTo = (double) $this->conn->fetchColumn($sql, $sqlParameters, 0);
+			$this->aggTo = (float) $this->conn->fetchColumn($sql, $sqlParameters, 0);
 		}
 
 		if (self::$debug) {
